+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/telqos/TriggeredValueMappingLookupCommand.java
@@ -5,9 +5,13 @@ import com.dwarfeng.fdr.stack.handler.TriggeredValueMappingLookupHandler;
 import com.dwarfeng.springtelqos.sdk.command.CliCommand;
 import com.dwarfeng.springtelqos.stack.command.Context;
 import com.dwarfeng.springtelqos.stack.exception.TelqosException;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
+import org.apache.commons.cli.ParseException;
 import org.apache.commons.lang3.tuple.Pair;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -16,12 +20,14 @@ import java.util.List;
 @Component
 public class TriggeredValueMappingLookupCommand extends CliCommand {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(TriggeredValueMappingLookupCommand.class);
+
     private static final String COMMAND_OPTION_LIST = "l";
     private static final String COMMAND_OPTION_CLEANUP = "c";
+    private static final String COMMAND_OPTION_LOOKUP_DETAIL = "ld";
 
     private static final String[] COMMAND_OPTION_ARRAY = new String[]{
-            COMMAND_OPTION_LIST,
-            COMMAND_OPTION_CLEANUP
+            COMMAND_OPTION_LIST, COMMAND_OPTION_CLEANUP, COMMAND_OPTION_LOOKUP_DETAIL
     };
 
     private static final String IDENTITY = "tml";
@@ -31,10 +37,11 @@ public class TriggeredValueMappingLookupCommand extends CliCommand {
             CommandUtil.concatOptionPrefix(COMMAND_OPTION_LIST);
     private static final String CMD_LINE_SYNTAX_CLEANUP = IDENTITY + " " +
             CommandUtil.concatOptionPrefix(COMMAND_OPTION_CLEANUP);
+    private static final String CMD_LINE_SYNTAX_LOOKUP_DETAIL = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_LOOKUP_DETAIL) + " session-id";
 
     private static final String[] CMD_LINE_ARRAY = new String[]{
-            CMD_LINE_SYNTAX_LIST,
-            CMD_LINE_SYNTAX_CLEANUP
+            CMD_LINE_SYNTAX_LIST, CMD_LINE_SYNTAX_CLEANUP, CMD_LINE_SYNTAX_LOOKUP_DETAIL
     };
 
     private static final String CMD_LINE_SYNTAX = CommandUtil.syntax(CMD_LINE_ARRAY);
@@ -46,11 +53,14 @@ public class TriggeredValueMappingLookupCommand extends CliCommand {
         this.handler = handler;
     }
 
+    @SuppressWarnings("DuplicatedCode")
     @Override
     protected List<Option> buildOptions() {
         List<Option> list = new ArrayList<>();
         list.add(Option.builder(COMMAND_OPTION_LIST).desc("查询会话信息").build());
         list.add(Option.builder(COMMAND_OPTION_CLEANUP).desc("清理会话信息").build());
+        list.add(Option.builder(COMMAND_OPTION_LOOKUP_DETAIL).desc("查询会话详细信息").hasArg(true).type(Number.class).
+                argName("session-id").build());
         return list;
     }
 
@@ -71,6 +81,9 @@ public class TriggeredValueMappingLookupCommand extends CliCommand {
                 case COMMAND_OPTION_CLEANUP:
                     handleCleanup(context);
                     break;
+                case COMMAND_OPTION_LOOKUP_DETAIL:
+                    handleLookupDetail(context, cmd);
+                    break;
             }
         } catch (Exception e) {
             throw new TelqosException(e);
@@ -79,12 +92,28 @@ public class TriggeredValueMappingLookupCommand extends CliCommand {
 
     private void handleList(Context context) throws Exception {
         List<MappingLookupSessionInfo> sessionInfos = handler.getSessionInfos();
-        MappingCommandUtil.printSessionInfos(context, sessionInfos);
+        CommandUtil.printSessionInfos(context, sessionInfos);
     }
 
     private void handleCleanup(Context context) throws Exception {
         handler.cleanup();
         context.sendMessage("清理完成!");
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    private void handleLookupDetail(Context context, CommandLine cmd) throws Exception {
+        long sessionId;
+        try {
+            sessionId = ((Number) cmd.getParsedOptionValue(COMMAND_OPTION_LOOKUP_DETAIL)).longValue();
+        } catch (ParseException e) {
+            LOGGER.warn("解析命令选项时发生异常，异常信息如下", e);
+            context.sendMessage("命令行格式错误，正确的格式为: " + CMD_LINE_SYNTAX_LOOKUP_DETAIL);
+            context.sendMessage("请留意选项 " + COMMAND_OPTION_LOOKUP_DETAIL + " 后接参数的类型应该是数字 ");
+            return;
+        }
+
+        MappingLookupSessionInfo sessionInfo = handler.getSessionInfo(new LongIdKey(sessionId));
+        CommandUtil.printSessionInfoDetail(context, sessionInfo);
     }
 
     @SuppressWarnings("DuplicatedCode")
@@ -98,6 +127,10 @@ public class TriggeredValueMappingLookupCommand extends CliCommand {
         if (cmd.hasOption(COMMAND_OPTION_CLEANUP)) {
             i++;
             subCmd = COMMAND_OPTION_CLEANUP;
+        }
+        if (cmd.hasOption(COMMAND_OPTION_LOOKUP_DETAIL)) {
+            i++;
+            subCmd = COMMAND_OPTION_LOOKUP_DETAIL;
         }
         return Pair.of(subCmd, i);
     }